Add get(fieldName) and set(fieldName, value) to ODocument
The old field() methods also calculate expressions, eg.

     doc.field(“foo.bar”)

returns the “bar” field of the “foo” field of the document.
This implies some of overhead and blocks us from having fields with
complex names (containing dots, spaces and other special characters).
For obvious backward compatibility reasons, the old API cannot be
changed, so the only solution is to introduce a new API (get, set in
this case)

Related to: #4236